+++ src/app/app.component.ts
@@ -1,15 +1,17 @@
 /*
- Copyright (c) 2025 gematik GmbH
- Licensed under the EUPL, Version 1.2 or - as soon they will be approved by
- the European Commission - subsequent versions of the EUPL (the "Licence");
- You may not use this work except in compliance with the Licence.
-    You may obtain a copy of the Licence at:
-    https://joinup.ec.europa.eu/software/page/eupl
-        Unless required by applicable law or agreed to in writing, software
- distributed under the Licence is distributed on an "AS IS" basis,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the Licence for the specific language governing permissions and
- limitations under the Licence.
+    Copyright (c) 2025 gematik GmbH
+    Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
+    European Commission – subsequent versions of the EUPL (the "Licence").
+    You may not use this work except in compliance with the Licence.
+    You find a copy of the Licence in the "Licence" file or at
+    https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the Licence is distributed on an "AS IS" basis,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
+    In case of changes by gematik find details in the "Readme" file.
+    See the Licence for the specific language governing permissions and limitations under the Licence.
+    *******
+    For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
  */
 
 import { Component, inject, OnDestroy, OnInit, signal, WritableSignal } from '@angular/core';
@@ -22,7 +24,7 @@ import { JwtHelperService } from '@auth0/angular-jwt';
 // don't shorten the import, or all tests will break
 import { AuthService } from './services/auth.service';
 import { NavigationEnd, Router } from '@angular/router';
-import { AppConstants } from './shared/app-constants';
+import { AppConstants, isNewDesignActivated } from './shared/app-constants';
 
 @Component({
   selector: 'app-root',
@@ -119,4 +121,5 @@ export class AppComponent implements OnInit, OnDestroy {
   }
 
   protected readonly environment = environment;
+  protected readonly isNewDesignActivated = isNewDesignActivated;
 }