+++ src/app/welcome-tile-new/welcome-tile-new.component.spec.ts
@@ -1,27 +1,39 @@
 /*
- Copyright (c) 2025 gematik GmbH
- Licensed under the EUPL, Version 1.2 or - as soon they will be approved by
- the European Commission - subsequent versions of the EUPL (the "Licence");
- You may not use this work except in compliance with the Licence.
-    You may obtain a copy of the Licence at:
-    https://joinup.ec.europa.eu/software/page/eupl
-        Unless required by applicable law or agreed to in writing, software
- distributed under the Licence is distributed on an "AS IS" basis,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the Licence for the specific language governing permissions and
- limitations under the Licence.
+    Copyright (c) 2025 gematik GmbH
+    Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
+    European Commission – subsequent versions of the EUPL (the "Licence").
+    You may not use this work except in compliance with the Licence.
+    You find a copy of the Licence in the "Licence" file or at
+    https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the Licence is distributed on an "AS IS" basis,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
+    In case of changes by gematik find details in the "Readme" file.
+    See the Licence for the specific language governing permissions and limitations under the Licence.
+    *******
+    For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
  */
 
-import { WelcomeTileNewComponent, WelcomeTileNewConfig } from './welcome-tile-new.component';
+import { WelcomeTileNewComponent } from './welcome-tile-new.component';
 import { MockBuilder, MockedComponentFixture, MockRender } from 'ng-mocks';
+import { WelcomeTileConfig } from '../welcome/welcome.component';
+import { Router } from '@angular/router';
+import { TestBed } from '@angular/core/testing';
 
 describe('WelcomeTileNewComponent', () => {
   let fixture: MockedComponentFixture<WelcomeTileNewComponent, Partial<WelcomeTileNewComponent>>;
   let component: WelcomeTileNewComponent;
 
-  beforeEach(() => MockBuilder(WelcomeTileNewComponent));
+  let router: Router;
 
-  const initConfig: WelcomeTileNewConfig = {
+  beforeEach(() => {
+    return MockBuilder(WelcomeTileNewComponent).provide({
+      provide: Router,
+      useValue: { navigateByUrl: jasmine.createSpy('navigateByUrl') },
+    });
+  });
+
+  const initConfig: WelcomeTileConfig = {
     id: 'test-id',
     titleTextRows: ['test-', 'title'],
     tooltip: 'test-tooltip',
@@ -37,6 +49,7 @@ describe('WelcomeTileNewComponent', () => {
   beforeEach(() => {
     fixture = MockRender(WelcomeTileNewComponent, { config: initConfig });
     component = fixture.point.componentInstance;
+    router = TestBed.inject(Router);
   });
 
   it('should create', () => {
@@ -51,5 +64,91 @@ describe('WelcomeTileNewComponent', () => {
     expect(component.config.logoImage).toBe(initConfig.logoImage);
     expect(component.config.contentParagraphs).toBe(initConfig.contentParagraphs);
     expect(component.config.buttonLabel).toBe(initConfig.buttonLabel);
+  });
+
+  describe('changeIsExpandedState', () => {
+    it('should toggle isExpanded state', () => {
+      expect(component.isExpanded()).toBeFalse();
+
+      component.changeIsExpandedState();
+      expect(component.isExpanded()).toBeTrue();
+
+      component.changeIsExpandedState();
+      expect(component.isExpanded()).toBeFalse();
+    });
+  });
+
+  describe('handleTileClick', () => {
+    it('should change expanded state when tile is expandable', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(true);
+      spyOn(component, 'changeIsExpandedState');
+
+      component.handleTileClick();
+
+      expect(component.changeIsExpandedState).toHaveBeenCalled();
+      expect(router.navigateByUrl).not.toHaveBeenCalled();
+    });
+
+    it('should navigate to destination when tile is not expandable', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(false);
+      component.config.destinationRouterLink = '/test-route';
+
+      component.handleTileClick();
+
+      expect(router.navigateByUrl).toHaveBeenCalledWith('/test-route');
+    });
+  });
+
+  describe('navigateTo', () => {
+    it('should navigate to the given destination', () => {
+      const destination = '/subtile-route';
+
+      component.navigateTo(destination);
+
+      expect(router.navigateByUrl).toHaveBeenCalledWith(destination);
+    });
+  });
+  describe('getIconName', () => {
+    it('should return chevron_right when tile is not expandable', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(false);
+
+      expect(component.getIconName()).toBe('chevron_right');
+    });
+
+    it('should return keyboard_arrow_up when tile is expandable and expanded', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(true);
+      component.isExpanded.set(true);
+
+      expect(component.getIconName()).toBe('keyboard_arrow_up');
+    });
+
+    it('should return keyboard_arrow_down when tile is expandable and not expanded', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(true);
+      component.isExpanded.set(false);
+
+      expect(component.getIconName()).toBe('keyboard_arrow_down');
+    });
+  });
+
+  describe('getIconType', () => {
+    it('should return chevron-right-logo when tile is not expandable', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(false);
+
+      expect(component.getIconType()).toBe('chevron-right-logo');
+    });
+
+    it('should return chevron-up-logo when tile is expandable and expanded', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(true);
+      component.isExpanded.set(true);
+
+      expect(component.getIconType()).toBe('chevron-up-logo');
+    });
+
+    it('should return chevron-down-logo when tile is expandable and not expanded', () => {
+      spyOn(component, 'isTileExpandable').and.returnValue(true);
+      component.isExpanded.set(false);
+
+      expect(component.getIconType()).toBe('chevron-down-logo');
+    });
   });
 });