+++ ReleaseNotes.md
@@ -2,6 +2,15 @@
 
 # Release portal-shell
 
-## Release 1.2.0
+## Release 1.2.3
+- Updated ospo-resources for adding additional notes and disclaimer
+- Added sub-tiles for non-nominal notification (FEATURE_FLAG_NON_NOMINAL_NOTIFICATION)
 
+## Release 1.2.2
+- Add new font and background color
+
+## Release 1.2.1
+- New wording for igs tile
+
+## Release 1.2.0
 - First official GitHub-Release